IN126049802/requirements.txt
langchain
langchain-community
langchain-core
langgraph
chromadb
pypdf
python-dotenv
groq
sentence-transformers